Start IOThrottlerPools with 1 second of bandwidth
There is a performance issue where the first item added to an 'IOThrottlerPool' will
have to wait 1 second to have any bandwidth. This is even true when a pool has
been constructed with 'Unlimited' bandwidth. By initializing the pool with one
seconds worth of bandwidth we eliminate the 1 second penalty without causing an
unreasonable bandwidth spike.