Only allocate half of what a client should use.
One more change to ensure equal distribution of bandwidth among clients in a
pool. If the pool size doesn't change and a client can use all it's alloted
bandwith then this change will double the number of times the client must get an
allocation of bandwidth, but it greatly improves the fairness.