Fix crash when pool has Unlimited bandwidth
We were casting items of type Bandwidth (int64) to an int. For Unlimited
bandwidth it would result in an -1 which is used to create a slice. We now cast
the read/write count to type Bandwidth. It's kind of weird to count byte counts
to Bandwidth but it's simpler than casting both the byte count and the Bandwidth
to int64.

Also added a unit test that would trigger the crash and also ensures that
reading/writing in a pool with Unlimited bandwidth is fast.